+++ SharedPackages/BrowserServicesKit/Sources/Subscription/Storage/V2/SubscriptionTokenKeychainStorageV2.swift
@@ -35,23 +35,42 @@ public enum KeychainErrorAuthVersion: String {
 
 public final class SubscriptionTokenKeychainStorageV2: AuthTokenStoring {
 
-    private let keychainType: KeychainType
     private let errorEventsHandler: (AccountKeychainAccessType, AccountKeychainAccessError) -> Void
     private let accessQueue = DispatchQueue(label: "keychain.subscription.access", qos: .userInitiated)
-    private let keychainOperations: KeychainOperationsProtocol
+    private let keychainManager: KeychainManager
 
     public init(keychainType: KeychainType = .dataProtection(.unspecified),
                 errorEventsHandler: @escaping (AccountKeychainAccessType, AccountKeychainAccessError) -> Void,
                 keychainOperations: KeychainOperationsProtocol = DefaultKeychainOperations()) {
-        self.keychainType = keychainType
         self.errorEventsHandler = errorEventsHandler
-        self.keychainOperations = keychainOperations
+        self.keychainManager = KeychainManager(keychainOperations: keychainOperations, attributes: Self.defaultAttributes(keychainType: keychainType))
+    }
+
+    /*
+     Uses just kSecAttrService as the primary key, since we don't want to store
+     multiple accounts/tokens at the same time
+     */
+    enum SubscriptionKeychainField: String, CaseIterable {
+        case tokenContainer = "subscription.v2.tokens"
+
+        var keyValue: String {
+            "com.duckduckgo" + "." + rawValue
+        }
+    }
+
+    private static func defaultAttributes(keychainType: KeychainType) -> [CFString: Any] {
+        var attributes: [CFString: Any] = [
+            kSecClass: kSecClassGenericPassword,
+            kSecAttrSynchronizable: false
+        ]
+        attributes.merge(keychainType.queryAttributes()) { $1 }
+        return attributes
     }
 
     public func getTokenContainer() throws -> TokenContainer? {
         return try accessQueue.sync {
             do {
-                guard let data = try retrieveData(forField: .tokenContainer) else {
+                guard let data = try keychainManager.retrieveData(forKey: SubscriptionKeychainField.tokenContainer.keyValue) else {
                     Logger.subscriptionKeychain.debug("TokenContainer not found")
                     return nil
                 }
@@ -73,7 +92,7 @@ public final class SubscriptionTokenKeychainStorageV2: AuthTokenStoring {
             do {
                 guard let tokenContainer else {
                     Logger.subscriptionKeychain.debug("Remove TokenContainer")
-                    try self.deleteItem(forField: .tokenContainer)
+                    try keychainManager.deleteItem(forKey: SubscriptionKeychainField.tokenContainer.keyValue)
                     return
                 }
 
@@ -81,7 +100,7 @@ public final class SubscriptionTokenKeychainStorageV2: AuthTokenStoring {
                     throw AccountKeychainAccessError.failedToEncodeKeychainData // Fixed error name
                 }
 
-                try self.store(data: data, forField: .tokenContainer)
+                try keychainManager.store(data: data, forKey: SubscriptionKeychainField.tokenContainer.keyValue)
             } catch {
                 Logger.subscriptionKeychain.fault("Failed to set TokenContainer: \(error, privacy: .public)")
                 if let error = error as? AccountKeychainAccessError {
@@ -93,111 +112,5 @@ public final class SubscriptionTokenKeychainStorageV2: AuthTokenStoring {
                 throw error
             }
         }
-    }
-}
-
-extension SubscriptionTokenKeychainStorageV2 {
-
-    /*
-     Uses just kSecAttrService as the primary key, since we don't want to store
-     multiple accounts/tokens at the same time
-     */
-    enum SubscriptionKeychainField: String, CaseIterable {
-        case tokenContainer = "subscription.v2.tokens"
-
-        var keyValue: String {
-            "com.duckduckgo" + "." + rawValue
-        }
-    }
-
-    func retrieveData(forField field: SubscriptionKeychainField) throws -> Data? {
-        var query = defaultAttributes()
-        query[kSecAttrService] = field.keyValue
-        query[kSecMatchLimit] = kSecMatchLimitOne
-        query[kSecReturnData] = true
-
-        var item: CFTypeRef?
-        let status = keychainOperations.copyMatching(query as CFDictionary, &item)
-
-        if status == errSecSuccess {
-            if let existingItem = item as? Data {
-                return existingItem
-            } else {
-                throw AccountKeychainAccessError.failedToDecodeKeychainData
-            }
-        } else if status == errSecItemNotFound {
-            return nil
-        } else {
-            throw AccountKeychainAccessError.keychainLookupFailure(status)
-        }
-    }
-
-    func store(data: Data, forField field: SubscriptionKeychainField) throws {
-        var query = defaultAttributes()
-        query[kSecAttrService] = field.keyValue
-        query[kSecAttrAccessible] = kSecAttrAccessibleAfterFirstUnlock
-        query[kSecValueData] = data
-
-        let status = keychainOperations.add(query as CFDictionary, nil)
-
-        switch status {
-        case errSecSuccess:
-            Logger.subscriptionKeychain.debug("Successfully added keychain item for \(field.keyValue)")
-            return
-        case errSecDuplicateItem:
-            Logger.subscriptionKeychain.debug("Keychain item exists, updating for \(field.keyValue)")
-            let updateStatus = updateData(data, forField: field)
-            guard updateStatus == errSecSuccess else {
-                Logger.subscriptionKeychain.error("Failed to update keychain item: \(updateStatus)")
-                throw AccountKeychainAccessError.keychainSaveFailure(updateStatus)
-            }
-            Logger.subscriptionKeychain.debug("Successfully updated keychain item for \(field.keyValue)")
-        default:
-            Logger.subscriptionKeychain.error("Failed to add keychain item: \(status.humanReadableDescription)")
-            throw AccountKeychainAccessError.keychainSaveFailure(status)
-        }
-    }
-
-    private func updateData(_ data: Data, forField field: SubscriptionKeychainField) -> OSStatus {
-        var query = defaultAttributes()
-        query[kSecAttrService] = field.keyValue
-
-        let newAttributes = [
-            kSecValueData: data,
-            kSecAttrAccessible: kSecAttrAccessibleAfterFirstUnlock
-        ] as [CFString: Any]
-
-        let status = keychainOperations.update(query as CFDictionary, newAttributes as CFDictionary)
-
-        if status != errSecSuccess {
-            Logger.subscriptionKeychain.error("SecItemUpdate failed with status: \(status.humanReadableDescription) for field: \(field.keyValue)")
-        }
-
-        return status
-    }
-
-    func deleteItem(forField field: SubscriptionKeychainField) throws {
-        var query = defaultAttributes()
-        query[kSecAttrService] = field.keyValue
-
-        let status = keychainOperations.delete(query as CFDictionary)
-
-        if status == errSecSuccess {
-            Logger.subscriptionKeychain.debug("Successfully deleted keychain item for \(field.keyValue)")
-        } else if status == errSecItemNotFound {
-            Logger.subscriptionKeychain.debug("Keychain item not found for deletion: \(field.keyValue)")
-        } else {
-            Logger.subscriptionKeychain.error("Failed to delete keychain item: \(status.humanReadableDescription)")
-            throw AccountKeychainAccessError.keychainDeleteFailure(status)
-        }
-    }
-
-    private func defaultAttributes() -> [CFString: Any] {
-        var attributes: [CFString: Any] = [
-            kSecClass: kSecClassGenericPassword,
-            kSecAttrSynchronizable: false
-        ]
-        attributes.merge(keychainType.queryAttributes()) { $1 }
-        return attributes
     }
 }
